♻️ refactor: 커리큘럼 편집 로직을 useCurriculumEditor 훅으로 추출
